Refactoring I: Remove PackagedOperation and block stuff
Remove all PackagedOperation related code and code for block structures
from linear_operator.h, remove corresponding tests and rename remaining
tests.
