+++ tests/lac/linear_operator_06.with_cxx11=on.output
@@ -1,25 +1,3 @@
 
-DEAL::MyMatrix1::vmult
-DEAL::MyMatrix1::Tvmult
-DEAL::MyMatrix1::vmult
-DEAL::MyMatrix1::Tvmult
-DEAL::MyMatrix2::vmult
-DEAL::MyMatrix2::Tvmult
-DEAL::MyMatrix2::vmult_add
-DEAL::MyMatrix2::Tvmult_add
-DEAL::MyMatrix3::vmult
-DEAL::MyMatrix3::Tvmult
-DEAL::MyMatrix3::vmult_add
-DEAL::MyMatrix3::Tvmult_add
-DEAL::MyMatrix4::vmult
-DEAL::MyMatrix4::Tvmult
-DEAL::MyMatrix4::vmult_add
-DEAL::MyMatrix4::Tvmult_add
-DEAL::MyMatrix5::vmult
-DEAL::MyMatrix5::Tvmult
-DEAL::MyMatrix5::vmult_add
-DEAL::MyMatrix5::Tvmult_add
-DEAL::MyMatrix6::vmult
-DEAL::MyMatrix6::Tvmult
-DEAL::MyMatrix6::vmult_add
-DEAL::MyMatrix6::Tvmult_add
+DEAL::OK
+DEAL::OK
