--- tests/lac/linear_operator_08.with_cxx11=on.output
@@ -1,3 +1,0 @@
-
-DEAL::OK
-DEAL::OK
