--- tests/lac/linear_operator_10.with_cxx11=on.output
@@ -1,2 +1,0 @@
-
-DEAL::OK
