--- tests/lac/linear_operator_11.with_cxx11=on.output
@@ -1,21 +1,0 @@
-
-DEAL:: -- vmult --
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.0    
-DEAL::[block 1 ]  3.0    
-DEAL::[block 2 ]  10.    
-DEAL:: -- vmult_add --
-DEAL::Block vector: v:
-DEAL::[block 0 ]  1.0    
-DEAL::[block 1 ]  4.0    
-DEAL::[block 2 ]  11.    
-DEAL:: -- Tvmult --
-DEAL::Block vector: u:
-DEAL::[block 0 ]  18.    
-DEAL::[block 1 ]  9.0    
-DEAL::[block 2 ]  0.0    
-DEAL:: -- Tvmult_add --
-DEAL::Block vector: u:
-DEAL::[block 0 ]  19.    
-DEAL::[block 1 ]  10.    
-DEAL::[block 2 ]  1.0
