--- tests/lac/linear_operator_13.with_cxx11=on.output
@@ -1,33 +1,0 @@
-
-DEAL::Block vector: v1:
-DEAL::[block 0 ]  0.0    
-DEAL::[block 1 ]  3.0    
-DEAL::[block 2 ]  14.    
-DEAL::Block vector: v2:
-DEAL::[block 0 ]  0.0    
-DEAL::[block 1 ]  3.0    
-DEAL::[block 2 ]  14.    
-DEAL::Block vector: v1:
-DEAL::[block 0 ]  18.    
-DEAL::[block 1 ]  15.    
-DEAL::[block 2 ]  0.0    
-DEAL::Block vector: v2:
-DEAL::[block 0 ]  18.    
-DEAL::[block 1 ]  15.    
-DEAL::[block 2 ]  0.0    
-DEAL::Block vector: v1:
-DEAL::[block 0 ]  1.0    
-DEAL::[block 1 ]  4.0    
-DEAL::[block 2 ]  15.    
-DEAL::Block vector: v2:
-DEAL::[block 0 ]  1.0    
-DEAL::[block 1 ]  4.0    
-DEAL::[block 2 ]  15.    
-DEAL::Block vector: v1:
-DEAL::[block 0 ]  19.    
-DEAL::[block 1 ]  16.    
-DEAL::[block 2 ]  1.0    
-DEAL::Block vector: v2:
-DEAL::[block 0 ]  19.    
-DEAL::[block 1 ]  16.    
-DEAL::[block 2 ]  1.0
