--- tests/lac/linear_operator_14.with_cxx11=on.output
@@ -1,40 +1,0 @@
-
-DEAL:: -- Matrix -- 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  1.00000000000 
-DEAL::[block 1 ]  10.0000000000 
-DEAL::[block 2 ]  10.0000000000 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.00000000000 
-DEAL::[block 1 ]  3.00000000000 
-DEAL::[block 2 ]  10.0000000000 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.00000000000 
-DEAL::[block 1 ]  0.00000000000 
-DEAL::[block 2 ]  5.00000000000 
-DEAL:: -- Inverse -- 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  1.00000000000 
-DEAL::[block 1 ]  -3.33333333333 
-DEAL::[block 2 ]  4.66666666667 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.00000000000 
-DEAL::[block 1 ]  0.333333333333 
-DEAL::[block 2 ]  -0.666666666667 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.00000000000 
-DEAL::[block 1 ]  0.00000000000 
-DEAL::[block 2 ]  0.200000000000 
-DEAL:: -- Identity -- 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  1.00000000000 
-DEAL::[block 1 ]  0.00000000000 
-DEAL::[block 2 ]  0.00000000000 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.00000000000 
-DEAL::[block 1 ]  1.00000000000 
-DEAL::[block 2 ]  0.00000000000 
-DEAL::Block vector: v:
-DEAL::[block 0 ]  0.00000000000 
-DEAL::[block 1 ]  0.00000000000 
-DEAL::[block 2 ]  1.00000000000
